+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/upgrade/v0_0_3/UpgradeJournalArticleType.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.journal.upgrade.v1_0_0;
+package com.liferay.journal.upgrade.v0_0_3;
 
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
 import com.liferay.journal.model.JournalArticle;
+import com.liferay.journal.upgrade.v1_0_0.UpgradeJournalArticles;
 import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
@@ -41,7 +42,6 @@ import com.liferay.portlet.asset.util.AssetVocabularySettingsHelper;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
